requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
neo4j==5.27.0
chromadb==0.5.23
google-generativeai==0.8.4
python-dotenv==1.0.1
streamlit==1.41.1
requests==2.32.3
orjson==3.10.13
streamlit-agraph==0.0.45
